+++ src/features/usage-consolidation-accordion/usage-consolidation-content-title.js
@@ -3,17 +3,15 @@ import PropTypes from 'prop-types';
 import {
   useIntl,
   FormattedMessage,
-  FormattedNumber,
 } from 'react-intl';
 import { Link } from 'react-router-dom';
-import getSymbolFromCurrency from 'currency-symbol-map';
 
 import {
-  MultiColumnList,
   NoValue,
   List,
 } from '@folio/stripes/components';
 
+import SummaryTable from './summary-table';
 import {
   formatCoverageYear,
   formatCoverageFullDate,
@@ -21,10 +19,9 @@ import {
   isBookPublicationType,
 } from '../../components/utilities';
 import {
-  formatCost,
-  formatValue,
-} from './utilities';
-import { costPerUse as costPerUseShape } from '../../constants';
+  costPerUse as costPerUseShape,
+  entityTypes,
+} from '../../constants';
 import styles from './usage-consolidation-content.css';
 
 const propTypes = {
@@ -33,19 +30,13 @@ const propTypes = {
   year: PropTypes.string.isRequired,
 };
 
-const UsageConsolidationContentTitle = ({
-  costPerUseData,
-  year,
-  publicationType,
-}) => {
+const UsageConsolidationContentTitle = (props) => {
   const intl = useIntl();
+  const { costPerUseData, publicationType } = props;
   const [sortedColumn, setSortedColumn] = useState('packageName');
   const [sortOrder, setSortOrder] = useState('ascending');
 
   const { data } = costPerUseData;
-
-  const currency = data?.attributes?.parameters?.currency;
-  const currencySymbol = getSymbolFromCurrency(currency) || '';
 
   const holdingsSummary = data?.attributes?.analysis?.holdingsSummary;
   const noCostPerUseAvailable = !holdingsSummary;
@@ -98,71 +89,59 @@ const UsageConsolidationContentTitle = ({
         />
       );
     },
-    cost: (rowData) => formatValue(rowData.cost, (value) => formatCost(value, currencySymbol, currency)),
-    costPerUse: (rowData) => formatValue(rowData.costPerUse, (value) => formatCost(value, currencySymbol, currency)),
-    usage: (rowData) => formatValue(rowData.usage, (usage) => <FormattedNumber value={usage} />),
   };
 
-  if (noCostPerUseAvailable) {
-    return (
-      <div data-test-usage-consolidation-error>
-        <FormattedMessage
-          id="ui-eholdings.usageConsolidation.summary.noData"
-          values={{
-            record: 'title',
-            year,
-          }}
-        />
-      </div>
-    );
-  }
+  const contentData = holdingsSummary ?
+    holdingsSummary.sort((a, b) => {
+      const valA = sortOrder === 'ascending' ? a : b;
+      const valB = sortOrder === 'ascending' ? b : a;
 
-  const sortedHoldingsSummary = holdingsSummary.sort((a, b) => {
-    const valA = sortOrder === 'ascending' ? a : b;
-    const valB = sortOrder === 'ascending' ? b : a;
-
-    if (sortedColumn === 'packageName') {
-      return valA.packageName.localeCompare(valB.packageName);
-    }
-    if (sortedColumn === 'coverage') {
-      const coverageA = [...valA.coverages].sort(compareCoveragesToBeSortedInDescOrder)[0];
-      const coverageB = [...valB.coverages].sort(compareCoveragesToBeSortedInDescOrder)[0];
-
-      // Move empty coverages to be first when ascending order, and last when descending order
-      if (!coverageA) {
-        return -1;
+      if (sortedColumn === 'packageName') {
+        return valA.packageName.localeCompare(valB.packageName);
       }
-      if (!coverageB) {
-        return 1;
+      if (sortedColumn === 'coverage') {
+        const coverageA = [...valA.coverages].sort(compareCoveragesToBeSortedInDescOrder)[0];
+        const coverageB = [...valB.coverages].sort(compareCoveragesToBeSortedInDescOrder)[0];
+
+        // Move empty coverages to be first when ascending order, and last when descending order
+        if (!coverageA) {
+          return -1;
+        }
+        if (!coverageB) {
+          return 1;
+        }
+
+        return compareCoveragesToBeSortedInDescOrder(coverageB, coverageA);
       }
 
-      return compareCoveragesToBeSortedInDescOrder(coverageB, coverageA);
-    }
+      return valA[sortedColumn] - valB[sortedColumn];
+    }) : [];
 
-    return valA[sortedColumn] - valB[sortedColumn];
-  });
+  const customProperties = {
+    visibleColumns: ['packageName', 'coverage', 'cost', 'usage', 'costPerUse'],
+    columnMapping: {
+      packageName: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.packageName' }),
+      coverage: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.coverage' }),
+      usage: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.title.usage' }),
+    },
+    columnWidths: {
+      packageName: '33%',
+      coverage: '33%',
+      cost: '12%',
+      usage: '10%',
+      costPerUse: '12%',
+    },
+    formatter,
+  };
 
   return (
-    <MultiColumnList
+    <SummaryTable
       id="titleUsageConsolidationSummary"
-      contentData={sortedHoldingsSummary}
-      visibleColumns={['packageName', 'coverage', 'cost', 'usage', 'costPerUse']}
-      formatter={formatter}
-      columnMapping={{
-        packageName: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.packageName' }),
-        coverage: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.coverage' }),
-        cost: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.packageCost' }),
-        usage: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.title.usage' }),
-        costPerUse: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.costPerUse' }),
-      }}
-      columnWidths={{
-        packageName: '33%',
-        coverage: '33%',
-        cost: '12%',
-        usage: '10%',
-        costPerUse: '12%',
-      }}
-      onHeaderClick={(e, metadata) => {
+      contentData={contentData}
+      entityType={entityTypes.TITLE}
+      customProperties={customProperties}
+      noCostPerUseAvailable={noCostPerUseAvailable}
+      onHeaderClick={(_, metadata) => {
         const { name } = metadata;
 
         if (name !== sortedColumn) {
@@ -175,6 +154,7 @@ const UsageConsolidationContentTitle = ({
       }}
       sortedColumn={sortedColumn}
       sortDirection={sortOrder}
+      {...props}
     />
   );
 };