+++ src/redux/actions/costperuse/get-cost-per-use-success.js
@@ -1,0 +1,6 @@
+export const GET_COST_PER_USE_SUCCESS = 'GET_COST_PER_USE_SUCCESS';
+
+export const getCostPerUseSuccess = payload => ({
+  type: GET_COST_PER_USE_SUCCESS,
+  payload,
+});