+++ src/redux/actions/costperuse/get-package-cost-per-use-success.js
@@ -1,6 +1,0 @@
-export const GET_PACKAGE_COST_PER_USE_SUCCESS = 'GET_PACKAGE_COST_PER_USE_SUCCESS';
-
-export const getPackageCostPerUseSuccess = payload => ({
-  type: GET_PACKAGE_COST_PER_USE_SUCCESS,
-  payload,
-});