+++ src/redux/epics/index.js
@@ -24,5 +24,4 @@ export { default as createGetUsageConsolidationEpic } from './get-usage-consolid
 export { default as createPostUsageConsolidationEpic } from './post-usage-consolidation';
 export { default as createPatchUsageConsolidationEpic } from './patch-usage-consolidation';
 export { default as createGetCurrenciesEpic } from './get-currencies';
-export { default as createGetPackageCostPerUseEpic } from './get-package-costperuse';
-export { default as createGetTitleCostPerUseEpic } from './get-title-costperuse';
+export { default as createGetCostPerUseEpic } from './get-costperuse';