+++ src/redux/reducers/costPerUse.js
@@ -1,10 +1,7 @@
 import {
-  GET_PACKAGE_COST_PER_USE,
-  GET_PACKAGE_COST_PER_USE_SUCCESS,
-  GET_PACKAGE_COST_PER_USE_FAILURE,
-  GET_TITLE_COST_PER_USE,
-  GET_TITLE_COST_PER_USE_SUCCESS,
-  GET_TITLE_COST_PER_USE_FAILURE,
+  GET_COST_PER_USE,
+  GET_COST_PER_USE_SUCCESS,
+  GET_COST_PER_USE_FAILURE,
 } from '../actions';
 
 import { formatErrors } from '../helpers';
@@ -33,12 +30,9 @@ const handlePendingRequest = (state) => ({
 });
 
 const handlers = {
-  [GET_PACKAGE_COST_PER_USE]: handlePendingRequest,
-  [GET_PACKAGE_COST_PER_USE_SUCCESS]: handleSuccess,
-  [GET_PACKAGE_COST_PER_USE_FAILURE]: handleError,
-  [GET_TITLE_COST_PER_USE]: handlePendingRequest,
-  [GET_TITLE_COST_PER_USE_SUCCESS]: handleSuccess,
-  [GET_TITLE_COST_PER_USE_FAILURE]: handleError,
+  [GET_COST_PER_USE]: handlePendingRequest,
+  [GET_COST_PER_USE_SUCCESS]: handleSuccess,
+  [GET_COST_PER_USE_FAILURE]: handleError,
 };
 
 const initialState = {