+++ test/bigtest/interactors/resource-usage-consolidation.js
@@ -6,12 +6,14 @@ import {
 import { AccordionInteractor } from '@folio/stripes-components/lib/Accordion/tests/interactor';
 
 import UsageConsolidationFilters from './usage-consolidation-filters';
+import UsageConsolidationContent from './usage-consolidation-content-resource';
 
 export default @interactor class ResourceUsageConsolidation {
   accordion = new AccordionInteractor('#resourceShowUsageConsolidation');
   isAccordionPresent = isPresent('#resourceShowUsageConsolidation');
 
   filters = new UsageConsolidationFilters();
+  content = new UsageConsolidationContent();
 
   whenLoaded() {
     return this.when(() => this.isAccordionPresent).timeout(1000);