+++ test/bigtest/tests/unit/reducers/cost-per-use-test.js
@@ -3,12 +3,9 @@ import { expect } from 'chai';
 
 import { costPerUse } from '../../../../../src/redux/reducers';
 import {
-  GET_PACKAGE_COST_PER_USE,
-  GET_PACKAGE_COST_PER_USE_SUCCESS,
-  GET_PACKAGE_COST_PER_USE_FAILURE,
-  GET_TITLE_COST_PER_USE,
-  GET_TITLE_COST_PER_USE_SUCCESS,
-  GET_TITLE_COST_PER_USE_FAILURE,
+  GET_COST_PER_USE,
+  GET_COST_PER_USE_SUCCESS,
+  GET_COST_PER_USE_FAILURE,
 } from '../../../../../src/redux/actions';
 
 describe('(reducer) costPerUse', () => {
@@ -21,13 +18,13 @@ describe('(reducer) costPerUse', () => {
     });
   });
 
-  it('should handle GET_PACKAGE_COST_PER_USE', () => {
+  it('should handle GET_COST_PER_USE', () => {
     const actualState = {
       data: {},
       errors: [],
       isLoading: false,
     };
-    const action = { type: GET_PACKAGE_COST_PER_USE };
+    const action = { type: GET_COST_PER_USE };
     const expectedState = {
       data: {},
       errors: [],
@@ -39,14 +36,14 @@ describe('(reducer) costPerUse', () => {
     expect(costPerUse(actualState, action)).to.deep.equal(expectedState);
   });
 
-  it('should handle GET_PACKAGE_COST_PER_USE_FAILURE', () => {
+  it('should handle GET_COST_PER_USE_FAILURE', () => {
     const actualState = {
       data: {},
       errors: [],
       isLoading: true,
     };
     const action = {
-      type: GET_PACKAGE_COST_PER_USE_FAILURE,
+      type: GET_COST_PER_USE_FAILURE,
       payload: { errors: { title: 'error' } },
     };
     const expectedState = {
@@ -60,74 +57,14 @@ describe('(reducer) costPerUse', () => {
     expect(costPerUse(actualState, action)).to.deep.equal(expectedState);
   });
 
-  it('should handle GET_PACKAGE_COST_PER_USE_SUCCESS', () => {
+  it('should handle GET_COST_PER_USE_SUCCESS', () => {
     const actualState = {
       data: {},
       errors: [],
       isLoading: true,
     };
     const action = {
-      type: GET_PACKAGE_COST_PER_USE_SUCCESS,
-      payload: { costPerUse: 0.4 },
-    };
-    const expectedState = {
-      data: { costPerUse: 0.4 },
-      isLoading: false,
-      isLoaded: true,
-      isFailed: false,
-      errors: [],
-    };
-
-    expect(costPerUse(actualState, action)).to.deep.equal(expectedState);
-  });
-
-  it('should handle GET_TITLE_COST_PER_USE', () => {
-    const actualState = {
-      data: {},
-      errors: [],
-      isLoading: false,
-    };
-    const action = { type: GET_TITLE_COST_PER_USE };
-    const expectedState = {
-      data: {},
-      errors: [],
-      isLoading: true,
-      isLoaded: false,
-      isFailed: false,
-    };
-
-    expect(costPerUse(actualState, action)).to.deep.equal(expectedState);
-  });
-
-  it('should handle GET_TITLE_COST_PER_USE_FAILURE', () => {
-    const actualState = {
-      data: {},
-      errors: [],
-      isLoading: true,
-    };
-    const action = {
-      type: GET_TITLE_COST_PER_USE_FAILURE,
-      payload: { errors: { title: 'error' } },
-    };
-    const expectedState = {
-      data: {},
-      isLoading: false,
-      isLoaded: false,
-      isFailed: true,
-      errors: [{ title: 'error' }],
-    };
-
-    expect(costPerUse(actualState, action)).to.deep.equal(expectedState);
-  });
-
-  it('should handle GET_TITLE_COST_PER_USE_SUCCESS', () => {
-    const actualState = {
-      data: {},
-      errors: [],
-      isLoading: true,
-    };
-    const action = {
-      type: GET_TITLE_COST_PER_USE_SUCCESS,
+      type: GET_COST_PER_USE_SUCCESS,
       payload: { costPerUse: 0.4 },
     };
     const expectedState = {